+++ src/renderer.ts
@@ -1,1 +1,0 @@
-export * from './providers/mygene/renderer.js';